# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: BidFlow API
  version: "1.0.0"
  description: |
    BidFlow is a bid/proposal management platform with ML-powered win probability prediction.
    All API routes are versioned under `/api/v1/`.

    ## Authentication
    Uses JWT tokens stored in httpOnly cookies (`access_token_cookie`, `refresh_token_cookie`).
    CSRF protection via double-submit cookie pattern (`X-CSRF-TOKEN` header).

    ## Roles
    - **Admin** — Full access to all resources
    - **Company** — Enquiry owner, can update bid status on own enquiries
    - **Sales Executive** — Internal bid management
    - **Bidder** — Submit bids on marketplace enquiries only

    ## Rate Limiting
    All endpoints are rate-limited. Limits are documented per-endpoint.
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT access token in httpOnly cookie
    headerCsrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF token from csrf_access_token cookie
  schemas:
...
    Enquiry:
      type: object
      properties:
        _id: { type: string }
        enquiryId: { type: string }
        customerName: { type: string }
        contactInformation: { type: string }
        productServiceRequired: { type: string }
        priority: { type: string, enum: [High, Medium, Low] }
        status: { type: string, enum: [Under Review, Quotation Prepared, Negotiation, Rejected, Order Received] }
        negotiable: { type: boolean }
        visibility: { type: string, enum: [public, internal] }
        industry: { type: string }
        tags: { type: array, items: { type: string } }
        bidCount: { type: integer }
        listingDeadline: { type: string, format: date-time }
        date: { type: string, format: date-time }
        notes: { type: string }
        createdBy: { type: string }
    Bid:
      type: object
      properties:
        _id: { type: string }
        bidId: { type: string }
        enquiryId: { type: string }
        amount: { type: number }
        currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD] }
        status: { type: string, enum: [Quotation Prepared, Under Review, Negotiation, Order Received, Rejected] }
        assignedEmployee: { type: string }
        industry: { type: string }
        submissionDate: { type: string }
        remarks: { type: string }
        tags: { type: array, items: { type: string } }
        teamSize: { type: integer }
        aiPrediction: { type: integer }
        shapExplanations: { type: array, items: { $ref: '#/components/schemas/SHAPExplanation' } }
        comments: { type: array, items: { $ref: '#/components/schemas/Comment' } }
        history: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } }
        createdBy: { type: string }
    Comment:
      type: object
      properties:
        _id: { type: string }
        text: { type: string }
        author: { type: string }
        userId: { type: string }
        date: { type: string, format: date-time }
    HistoryEntry:
      type: object
      properties:
        status: { type: string }
        date: { type: string, format: date-time }
        note: { type: string }
    SHAPExplanation:
      type: object
      properties:
        feature: { type: string }
        value: { type: string }
        shap_value: { type: number }
        impact: { type: string, enum: [positive, negative] }
        text: { type: string }
...
    Document:
      type: object
      properties:
        _id: { type: string }
        filename: { type: string }
        path: { type: string }
        bidId: { type: string }
        enquiryId: { type: string }
        uploadedBy: { type: string }
        uploadDate: { type: string, format: date-time }
...
paths:
...
  /marketplace/:
    get:
      tags: [Marketplace]
      summary: List public enquiries
      description: Sealed bids — bidders cannot see each other's bids. Rate limited to 60/min.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: industry
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string, enum: [High, Medium, Low] }
        - name: sort
          in: query
          schema: { type: string, enum: [date, -date, priority, -bidCount], default: -date }
      responses:
        200:
          description: Paginated marketplace listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Enquiry' } }
                  page: { type: integer }
                  size: { type: integer }
                  total: { type: integer }
  /marketplace/{enquiry_id}:
    get:
      tags: [Marketplace]
      summary: Get marketplace enquiry detail
      description: Returns enquiry, documents, my bids (bidders), all bids (company/admin).
      security: [{ cookieAuth: [] }]
      parameters:
        - name: enquiry_id
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Enquiry detail with bids and documents
        404:
          description: Not found
  /marketplace/{enquiry_id}/bid:
    post:
      tags: [Marketplace]
      summary: Submit sealed bid
      description: Only Bidder, Sales Executive, or Company roles. Deadline enforced. Rate limited to 10/min.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: enquiry_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, minimum: 0 }
                currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD], default: USD }
                industry: { type: string }
                submissionDate: { type: string }
                remarks: { type: string }
                tags: { type: array, items: { type: string } }
                teamSize: { type: integer, minimum: 1 }
      responses:
        201:
          description: Bid submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg: { type: string }
                  bid: { $ref: '#/components/schemas/Bid' }
        400:
          description: Validation error or deadline passed
        403:
          description: Forbidden

  # ── Documents ────────────────────────────────────────────────────────────────
...
  /search:
    get:
      tags: [Search]
      summary: Global search
      description: Searches enquiries, bids, and documents. Role-scoped. Max 200 chars. Rate limited to 30/min.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, maxLength: 200 }
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  enquiries: { type: array, items: { $ref: '#/components/schemas/Enquiry' } }
                  bids: { type: array, items: { $ref: '#/components/schemas/Bid' } }
                  documents: { type: array, items: { $ref: '#/components/schemas/Document' } }

  # ── Tags ─────────────────────────────────────────────────────────────────────
  /tags/:
    get:
      tags: [Tags]
      summary: Get unique tags
      description: Returns sorted unique tags from enquiries and bids visible to the user.
      security: [{ cookieAuth: [] }]
      responses:
        200:
          description: Tag list
          content:
            application/json:
              schema:
                type: array
                items: { type: string }

  # ── Health (top-level, no /api/v1 prefix) ───────────────────────────────────
...